+++ src/frontend/apps/e2e/__tests__/app-impress/doc-export.spec.ts
@@ -298,7 +298,7 @@ test.describe('Doc Export', () => {
       })
       .click();
 
-    await page
+    void page
       .getByRole('button', {
         name: 'Download',
       })