+++ src/frontend/apps/impress/src/tests/utils.tsx
@@ -1,3 +1,4 @@
+import { TreeProvider } from '@gouvfr-lasuite/ui-kit';
 import { CunninghamProvider } from '@openfun/cunningham-react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { PropsWithChildren } from 'react';
@@ -14,8 +15,10 @@ export const AppWrapper = ({ children }: PropsWithChildren) => {
   });
 
   return (
-    <QueryClientProvider client={queryClient}>
-      <CunninghamProvider theme="default">{children}</CunninghamProvider>
-    </QueryClientProvider>
+    <TreeProvider initialTreeData={[]}>
+      <QueryClientProvider client={queryClient}>
+        <CunninghamProvider theme="default">{children}</CunninghamProvider>
+      </QueryClientProvider>
+    </TreeProvider>
   );
 };